agents/data_analyst.py
from typing import Dict, Any, List
from .base_agent import TaktAgent
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score

class DataAnalyst(TaktAgent):
    def analyze(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """
        Performs data analysis and predictive modeling for TAKT optimization
        """
        analysis = self.chain_of_thought(context)
        
        reasoning_steps = [
            self._collect_historical_data(context),
            self._predict_performance(context),
            self._identify_risks(context),
            self._optimize_resources(context),
            self._automate_decisions(context),
            self._compare_scenarios(context)
        ]
        
        analysis['reasoning_steps'] = reasoning_steps
        analysis['conclusions'] = self._generate_conclusions(reasoning_steps)
        analysis['recommendations'] = self._generate_recommendations(reasoning_steps)
        
        return analysis

    def validate(self, analysis: Dict[str, Any]) -> bool:
        """
        Validates data analysis and predictions
        """
        validation_criteria = [
            self._validate_data_quality(analysis),
            self._validate_model_performance(analysis),
            self._validate_predictions(analysis)
        ]
        
        return all(validation_criteria)

    def _collect_historical_data(self, context: Dict[str, Any]) -> Dict[str, Any]:
        return {
            'step': 'Historical Data Collection',
            'analysis': {
                'project_data': self._gather_project_data(context),
                'performance_metrics': self._collect_metrics(context),
                'patterns': self._identify_patterns(context)
            }
        }

    def _predict_performance(self, context: Dict[str, Any]) -> Dict[str, Any]:
        return {
            'step': 'Performance Prediction',
            'analysis': {
                'model_results': self._train_predict_model(context),
                'accuracy_metrics': self._evaluate_accuracy(context),
                'predictions': self._generate_predictions(context)
            }
        }

    def _train_predict_model(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """
        Train and evaluate a predictive model for TAKT performance
        """
        # Get historical data
        data = self._prepare_training_data(context)
        
        if data is None or len(data) == 0:
            return {'error': 'Insufficient data for model training'}

        # Split features and target
        X = data['features']
        y = data['target']

        # Split data into training and testing sets
        X_train, X_test, y_train, y_test = train_test_split(
            X, y, test_size=0.2, random_state=42
        )

        # Train model
        model = RandomForestRegressor(n_estimators=100, random_state=42)
        model.fit(X_train, y_train)

        # Make predictions
        y_pred = model.predict(X_test)

        # Evaluate model
        metrics = {
            'mse': mean_squared_error(y_test, y_pred),
            'r2': r2_score(y_test, y_pred),
            'feature_importance': dict(zip(data['feature_names'], 
                                        model.feature_importances_))
        }

        return {
            'model': model,
            'metrics': metrics,
            'predictions': y_pred.tolist()
        }

    # Implementation of remaining methods
    def _identify_risks(self, context: Dict[str, Any]) -> Dict[str, Any]:
        pass

    def _optimize_resources(self, context: Dict[str, Any]) -> Dict[str, Any]:
        pass

    def _automate_decisions(self, context: Dict[str, Any]) -> Dict[str, Any]:
        pass

    def _compare_scenarios(self, context: Dict[str, Any]) -> Dict[str, Any]:
        pass

    def _generate_conclusions(self, reasoning_steps: list) -> list:
        pass

    def _generate_recommendations(self, reasoning_steps: list) -> list:
        pass

    # Validation helper methods
    def _validate_data_quality(self, analysis: Dict[str, Any]) -> bool:
        pass

    def _validate_model_performance(self, analysis: Dict[str, Any]) -> bool:
        pass

    def _validate_predictions(self, analysis: Dict[str, Any]) -> bool:
        pass

    def _prepare_training_data(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """Prepare data for model training"""
        pass
